+++ demo/src/main/java/com/example/demo/controller/ModelController.java
@@ -3,15 +3,13 @@ package com.example.demo.controller;
 import com.example.demo.model.Product;
 import com.example.demo.service.ModelService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/modelList")
+@CrossOrigin(origins = "http://localhost:3000")
 public class ModelController {
 
     private final ModelService modelService;
