+++ demo/src/main/java/com/example/demo/controller/UserController.java
@@ -3,14 +3,15 @@ package com.example.demo.controller;
 import com.example.demo.model.User;
 import com.example.demo.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+
 
 @RestController
-@RequestMapping("/users")
+@RequestMapping("/user")
+@CrossOrigin(origins = "http://localhost:3000")
 public class UserController {
 
     private final UserService userService;
@@ -20,8 +21,46 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping
+    @GetMapping("/test")
     public List<User> getAllUsers() {
         return userService.getAllUsers();
     }
+
+    @PostMapping("/login")
+    public int login(@RequestBody Map<String, String> userInfo) {
+        String id = userInfo.get("userId");
+        String pw = userInfo.get("passWord");
+        System.out.println("login user: " + id);
+        System.out.println("login pw: " + pw);
+
+        System.out.println("result login: " + userService.login(id,pw));
+
+        return userService.login(id,pw);
+    }
+
+    @PostMapping("/newUser")
+    public int newUser(@RequestBody Map<String, String> newUserInfo) {
+        String newId = newUserInfo.get("userId");
+        String newPw = newUserInfo.get("passWord");
+        String newName = newUserInfo.get("userName");
+        System.out.println("newUserInfo user: " + newId);
+        System.out.println("newUserInfo pw: " + newPw);
+        System.out.println("newUserInfo name: " + newName);
+
+        User newUser = new User();
+        newUser.setUserid(newId);
+        newUser.setPassword(newPw);
+        newUser.setUsername(newName);
+
+        //중복된 계정이 있는지 확인 후 생성
+        int checkDuplicate = userService.duplicateCheckUser(newId);
+        if(checkDuplicate == 1){
+            System.out.println("중복 들어옴");
+            return 0;
+        }
+
+        return userService.newUser(newUser);
+    }
+
+
 }
