+++ demo/src/main/java/com/example/demo/mapper/UserMapper.java
@@ -1,11 +1,11 @@
 package com.example.demo.mapper;
 
-import com.example.demo.model.User;
+import com.example.demo.model.UserMb;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface UserMapper {
-    List<User> findAll();
+    List<UserMb> findAll();
 }
