+++ demo/src/main/java/com/example/demo/model/Product.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 import java.sql.Date;
 
 @Entity
-@Table(name = "Product")
+@Table(name = "products")
 public class Product {
 
     @Id
