+++ demo/src/main/java/com/example/demo/model/User.java
@@ -1,10 +1,47 @@
 package com.example.demo.model;
 
-import lombok.Data;
+import jakarta.persistence.*;
 
-@Data
+@Entity
+@Table(name = "users")
 public class User {
-    private Long id;
+
+    @Id
+    private String userid;
     private String username;
-    private String email;
+    private String password;
+
+    public enum Status {
+        ACTIVE, INACTIVE, SUSPENDED
+    }
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Status status = Status.ACTIVE; // 기본 값을 설정
+
+    // Getters and Setters
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getUserid() {
+        return userid;
+    }
+
+    public void setUserid(String userid) {
+        this.userid = userid;
+    }
+
 }
