--- demo/src/main/java/com/example/demo/model/UserJpa.java
@@ -1,47 +1,0 @@
-package com.example.demo.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "users")
-public class UserJpa {
-
-    @Id
-    private String userid;
-    private String username;
-    private String password;
-
-    public enum Status {
-        ACTIVE, INACTIVE, SUSPENDED
-    }
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Status status = Status.ACTIVE; // 기본 값을 설정
-
-    // Getters and Setters
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getUserid() {
-        return userid;
-    }
-
-    public void setUserid(String userid) {
-        this.userid = userid;
-    }
-
-}
