+++ demo/src/main/java/com/example/demo/model/UserMb.java
@@ -1,0 +1,10 @@
+package com.example.demo.model;
+
+import lombok.Data;
+
+@Data
+public class UserMb {
+    private Long id;
+    private String username;
+    private String email;
+}
