+++ demo/src/main/java/com/example/demo/service/UserMbService.java
@@ -1,0 +1,8 @@
+package com.example.demo.service;
+
+import com.example.demo.model.UserMb;
+import java.util.List;
+
+public interface UserMbService {
+    List<UserMb> getAllUsers();
+}
