+++ demo/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.datasource.username=root
 spring.datasource.password=0000
 
 # JPA
-spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
