+++ demo/src/main/resources/mapper/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.example.demo.mapper.UserMapper">
 
-    <select id="findAll" resultType="com.example.demo.model.User">
+    <select id="findAll" resultType="com.example.demo.model.UserMb">
         SELECT * FROM users
     </select>
